#43 新規登録・ログイン画面のパフォーマンス修正
