--- src/app/(authenticate)/login/page.tsx
@@ -1,105 +1,0 @@
-"use client"
-
-import { useActionState, useEffect } from 'react';
-import Link from 'next/link';
-import { submitSignInForm } from '@/lib/actions/signIn';
-import { useSession } from 'next-auth/react';
-import { useRouter } from 'next/navigation';
-
-export default function SignInPage() {
-  const [state, formAction, isPending] = useActionState(
-    submitSignInForm,
-    undefined,
-  );
-  const { update } = useSession();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (state?.success) {
-      update().then(() => {
-        router.refresh();
-        router.push('/');
-      });
-    }
-  }, [state]);
-
-  return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-      <div className="max-w-md w-full space-y-8">
-        <div>
-          <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
-            アカウントにログイン
-          </h2>
-          <p className="mt-2 text-center text-sm text-gray-600">
-            または{' '}
-            <Link
-              href="/auth/signup"
-              className="font-medium text-indigo-600 hover:text-indigo-500"
-            >
-              新しいアカウントを作成
-            </Link>
-          </p>
-        </div>
-        <form className="mt-8 space-y-6" action={formAction}>
-          <div className="rounded-md shadow-sm -space-y-px">
-            <div>
-              <label htmlFor="email" className="sr-only">
-                メールアドレス
-              </label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                autoComplete="email"
-                required
-                defaultValue={state?.values?.email || ''}
-                className={`appearance-none rounded-none relative block w-full px-3 py-2 border placeholder-gray-500 text-gray-900 rounded-t-md focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm ${
-                  state?.errors?.email ? 'border-red-500' : 'border-gray-300'
-                }`}
-                placeholder="メールアドレス"
-              />
-              {state?.errors?.email && (
-                <p className="mt-1 text-sm text-red-600">{state.errors.email}</p>
-              )}
-            </div>
-            <div>
-              <label htmlFor="password" className="sr-only">
-                パスワード
-              </label>
-              <input
-                id="password"
-                name="password"
-                type="password"
-                autoComplete="current-password"
-                required
-                className={`appearance-none rounded-none relative block w-full px-3 py-2 border placeholder-gray-500 text-gray-900 rounded-b-md focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm ${
-                  state?.errors?.password ? 'border-red-500' : 'border-gray-300'
-                }`}
-                placeholder="パスワード"
-              />
-              {state?.errors?.password && (
-                <p className="mt-1 text-sm text-red-600">{state.errors.password}</p>
-              )}
-            </div>
-          </div>
-
-          {state?.errors?.commom && (
-            <div className="rounded-md bg-red-50 p-4">
-              <div className="text-sm text-red-700">{state.errors.commom}</div>
-            </div>
-          )}
-
-          <div>
-            <button
-              type="submit"
-              disabled={isPending}
-              className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {isPending ? 'ログイン中...' : 'ログイン'}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
-}
