--- src/app/(authenticate)/signup/page.tsx
@@ -1,107 +1,0 @@
-"use client"
-
-import { useActionState } from 'react';
-import Link from 'next/link';
-import { submitSignUpForm } from '@/lib/actions/signUp';
-
-export default function SignUpPage() {
-  const [state, formAction, isPending] = useActionState(
-    submitSignUpForm,
-    undefined,
-  );
-
-  return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-      <div className="max-w-md w-full space-y-8">
-        <div>
-          <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
-            アカウントを作成
-          </h2>
-          <p className="mt-2 text-center text-sm text-gray-600">
-            すでにアカウントをお持ちの方は{' '}
-            <Link
-              href="/login"
-              className="font-medium text-indigo-600 hover:text-indigo-500"
-            >
-              ログイン
-            </Link>
-          </p>
-        </div>
-
-        <form className="mt-8 space-y-6" action={formAction}>
-          <div className="rounded-md shadow-sm -space-y-px">
-            <div>
-              <label htmlFor="name" className="sr-only">
-                名前
-              </label>
-              <input
-                id="name"
-                name="name"
-                type="text"
-                autoComplete="name"
-                required
-                defaultValue={state?.values?.name || ''}
-                className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-t-md focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm"
-                placeholder="名前"
-              />
-              {state?.errors?.name && (
-                <p className="mt-1 text-sm text-red-600">{state.errors.name}</p>
-              )}
-            </div>
-            <div>
-              <label htmlFor="email" className="sr-only">
-                メールアドレス
-              </label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                autoComplete="email"
-                required
-                defaultValue={state?.values?.email || ''}
-                className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm"
-                placeholder="メールアドレス"
-              />
-              {state?.errors?.email && (
-                <p className="mt-1 text-sm text-red-600">{state.errors.email}</p>
-              )}
-            </div>
-            <div>
-              <label htmlFor="password" className="sr-only">
-                パスワード
-              </label>
-              <input
-                id="password"
-                name="password"
-                type="password"
-                autoComplete="new-password"
-                required
-                className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-b-md focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm"
-                placeholder="パスワード"
-              />
-              {state?.errors?.password && (
-                <p className="mt-1 text-sm text-red-600">{state.errors.password}</p>
-              )}
-            </div>
-          </div>
-
-          {state?.errors?.commom && (
-            <div className="rounded-md bg-red-50 p-4">
-              <div className="text-sm text-red-700">{state.errors.commom}</div>
-            </div>
-          )}
-
-          <div>
-            <button
-              type="submit"
-              disabled={isPending}
-              className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {isPending ? '登録中...' : '新規登録'}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
-}
