+++ src/app/(public)/signin/page.tsx
@@ -1,0 +1,28 @@
+import Link from 'next/link';
+import LoginForm from './SignInForm';
+
+export default function SignInPage() {
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
+      <div className="max-w-md w-full space-y-8">
+        <div>
+          <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
+            アカウントにログイン
+          </h2>
+          <p className="mt-2 text-center text-sm text-gray-600">
+            または{' '}
+            <Link
+              href="/auth/signup"
+              className="font-medium text-indigo-600 hover:text-indigo-500"
+            >
+              新しいアカウントを作成
+            </Link>
+          </p>
+        </div>
+
+        {/* ログイン入力フォーム */}
+        <LoginForm />
+      </div>
+    </div>
+  );
+}
